Create a new class for daily weather
An array was added to the Weather object, where the weather objects for the week were added.

A new object was created for the daily weather using a class. The object contains the date, weather icon, and minimum and maximum temperatures in both Celsius and Fahrenheit.

The Day object is added using the Weather object's method.