Create a function that displays the weather for the week
For the weather of the week, a card had to be created for each day to contain the content. To do this, the weather icon insertion function had to be modified slightly so that it could also be used for days of the week.

The date-fns package was downloaded for date formatting.